refactor: 매번 캔버스를 만들지 않기 위해서 구조 변경 캔버스 만드는 부분을 field로 올림 thanos 캔버스를 만들어놓고 endRound에만 타노스를 그리게한다
